+++ api/controllers/silence/active.js
@@ -1,4 +1,4 @@
-/* global sails, Status, Meta, Logs, Songs */
+/* global sails, Status, Meta, Logs, Songs, Promise, Announcements, moment */
 
 module.exports = {
 
@@ -31,7 +31,7 @@ module.exports = {
                 await Logs.create({attendanceID: Meta['A'].attendanceID, logtype: 'silence-track', loglevel: 'warning', logsubtype: Meta['A'].show, event: `Track skipped due to silence.<br />Track: ${Meta.automation[0].ID} (${Meta.automation[0].Artist} - ${Meta.automation[0].Title})`})
                         .tolerate((err) => {
                         });
-                
+
                 // Skip the track if there's a track playing in automation
                 if (typeof Meta.automation[1] !== 'undefined')
                 {
@@ -40,6 +40,39 @@ module.exports = {
                     sails.helpers.rest.cmd('StopPlayer', 0);
                 }
             }
+
+            // If we are in automation, and prevError is less than 3 minutes ago, assume an audio issue and switch RadioDJs
+            if (moment().isAfter(moment(Status.errorCheck.prevError).add(3, 'minutes')) && (Meta['A'].state.startsWith("automation_") || Meta['A'].state === 'live_prerecord'))
+            {
+                await Meta.changeMeta({changingState: `Switching automation instances due to no audio`});
+                await Logs.create({attendanceID: Meta['A'].attendanceID, logtype: 'system', loglevel: 'danger', logsubtype: '', event: `Switching automation instances; silence detection executed multiple times.`})
+                        .tolerate((err) => {
+                            sails.log.error(err);
+                        });
+                await Announcements.findOrCreate({type: 'djcontrols', title: `Audio Error (system)`, announcement: "System recently had switched automation instances because the silence detectionm system triggered multiple times. Please check the logs for more info."}, {type: 'djcontrols', level: 'urgent', title: `Audio Error (system)`, announcement: "System recently had switched automation instances because the silence detectionm system triggered multiple times. Please check the logs for more info.", starts: moment().toISOString(true), expires: moment({year: 3000}).toISOString(true)})
+                        .tolerate((err) => {
+                            sails.log.error(err);
+                        });
+                var maps = sails.config.custom.radiodjs
+                        .filter((instance) => instance.rest === Meta['A'].radiodj)
+                        .map(async (instance) => {
+                            var status = await Status.findOne({name: `radiodj-${instance.name}`});
+                            if (status && status.status !== 1)
+                                await Status.changeStatus([{name: `radiodj-${instance.name}`, label: `RadioDJ ${instance.label}`, status: 2, data: `RadioDJ triggered queueFrozen multiple times; it has probably crashed.`}]);
+                            return true;
+                        });
+                await Promise.all(maps);
+                sails.sockets.broadcast('system-error', 'system-error', true);
+                await sails.helpers.rest.cmd('EnableAutoDJ', 0, 0);
+                await sails.helpers.rest.cmd('EnableAssisted', 1, 0);
+                await sails.helpers.rest.cmd('StopPlayer', 0, 0);
+                await sails.helpers.rest.changeRadioDj();
+                await sails.helpers.rest.cmd('ClearPlaylist', 1);
+                await sails.helpers.error.post();
+                await Meta.changeMeta({changingState: null});
+            }
+
+            Status.errorCheck.prevError = moment();
             return exits.success();
         } catch (e) {
             return exits.error(e);